refactor(api): split index DB modules
